backend/restapi/serializers.py
from rest_framework import serializers
from django.db import models
from .models import Product


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model=Product
        fields='__all__'
